// src/components/form-dynamic/models/field.enum.ts
export enum EValidation {
    required = 'required',
    email = 'email',
    maxLentgth = 'maxLentgth',
    minLentgth = 'minLentgth',
    onlyNumbers = 'onlyNumbers',
    onlyText = 'onlyText',
    compare = 'compare'
}

export enum ETypeField {
    text = 'text',
    email = 'email',
    password = 'password',
    date = 'date',
}